+++ fixtures/eslint/.eslintrc.json
@@ -10,6 +10,6 @@
   "plugins": ["react-hooks"],
   "rules": {
     "react-hooks/rules-of-hooks": 2,
-    "react-hooks/reactive-deps": 2
+    "react-hooks/exhaustive-deps": 2
   }
 }
